Added a "side" field to the Order class to determine buy/sell. Refactored the ID field in the Order class to Long. Modified test scripts (Each client generates 10 random orders).

+++ src/OrderRouter/Router.java
@@ -8,10 +8,10 @@ import Ref.Instrument;
 public interface Router {
     enum api {routeOrder, sendCancel, priceAtSize}
 
-    void routeOrder(int id, int sliceId, int size, Instrument i) throws IOException;
+    void routeOrder(long id, int sliceId, int size, Instrument i) throws IOException;
 
-    void sendCancel(int id, int sliceId, int size, Instrument i);
+    void sendCancel(long id, int sliceId, int size, Instrument i);
 
-    void priceAtSize(int id, int sliceId, Instrument i, int size) throws IOException;
+    void priceAtSize(long id, int sliceId, Instrument i, int size) throws IOException;
 
 }
